Compute k2ell in int32 instead of int16
Solving issue #28

+++ lensit/ffs_covs/ell_mat.py
@@ -77,7 +77,7 @@ class ell_mat:
             :math:`\ell = \rm{int}\left(|k| - \frac 12 \right)`
 
         """
-        ret = np.uint16(np.round(k - 0.5) + 0.5 * ((k - 0.5) < 0))
+        ret = np.uint32(np.round(k - 0.5) + 0.5 * ((k - 0.5) < 0))
         return ret
 
     def __call__(self, *args, **kwargs):
